+++ src/app/http/controllers/OrderController.php
@@ -1,8 +1,11 @@
 <?php namespace Redooor\Redminportal\App\Http\Controllers;
 
+use Exception;
 use Redooor\Redminportal\App\Models\User;
 use Redooor\Redminportal\App\Models\Order;
 use Redooor\Redminportal\App\Models\Product;
+use Redooor\Redminportal\App\Models\Bundle;
+use Redooor\Redminportal\App\Models\Coupon;
 
 class OrderController extends Controller
 {
@@ -23,6 +26,8 @@ class OrderController extends Controller
     public function getCreate()
     {
         $products = Product::orderBy('name')->lists('name', 'id');
+        $bundles = Bundle::orderBy('name')->lists('name', 'id');
+        $coupons = Coupon::orderBy('code')->lists('code', 'id');
 
         $payment_statuses = array(
             'Completed'     => 'Completed',
@@ -34,6 +39,8 @@ class OrderController extends Controller
 
         return view('redminportal::orders/create')
             ->with('products', $products)
+            ->with('bundles', $bundles)
+            ->with('coupons', $coupons)
             ->with('payment_statuses', $payment_statuses);
     }
     
@@ -53,7 +60,8 @@ class OrderController extends Controller
         $sid = \Input::get('id');
 
         $rules = array(
-            'product_id'        => 'required',
+            'product_id'        => 'required_without:bundle_id',
+            'bundle_id'         => 'required_without:product_id',
             'transaction_id'    => 'required',
             'payment_status'    => 'required',
             'paid'              => 'numeric',
@@ -73,21 +81,33 @@ class OrderController extends Controller
         $paid           = \Input::get('paid');
         $email          = \Input::get('email');
         
-        $save_to_order = array();
+        $apply_to_models = array();
+        
+        // Save products to order
         $products = \Input::get('product_id');
         if (count($products) > 0) {
             foreach ($products as $item) {
                 $model = Product::find($item);
                 if ($model != null) {
-                    $save_to_order[] = $model;
+                    $apply_to_models[] = $model;
+                }
+            }
+        }
+        // Save bundles to order
+        $bundles = \Input::get('bundle_id');
+        if (count($bundles) > 0) {
+            foreach ($bundles as $item) {
+                $model = Bundle::find($item);
+                if ($model != null) {
+                    $apply_to_models[] = $model;
                 }
             }
         }
 
-        // No product to add
-        if (count($save_to_order) == 0) {
+        // No product/bundle to add
+        if (count($apply_to_models) == 0) {
             $errors = new \Illuminate\Support\MessageBag;
-            $errors->add('productError', "The Products selected may have been deleted. Please try again.");
+            $errors->add('productError', "The items selected may have been deleted. Please try again.");
             return redirect($redirect_url)->withErrors($errors)->withInput();
         }
         
@@ -107,9 +127,35 @@ class OrderController extends Controller
         $new_order->payment_status = $payment_status;
         $new_order->save();
         
-        // Save the products
-        foreach ($save_to_order as $product) {
-            $new_order->products()->save($product);
+        // Save the products/bundles
+        foreach ($apply_to_models as $apply_to_model) {
+            $apply_to_model->orders()->save($new_order);
+        }
+        
+        // Save coupons to order
+        $coupons = \Input::get('coupon_id');
+        if (count($coupons) > 0) {
+            $errors = new \Illuminate\Support\MessageBag;
+            foreach ($coupons as $item) {
+                $model = Coupon::find($item);
+                if ($model != null) {
+                    try {
+                        $new_order->addCoupon($model);
+                    } catch (Exception $exp) {
+                        $errors->add('couponError',
+                            "Coupon " . $model->code . " cannot be added because: " . $exp->getMessage()
+                        );
+                        
+                    }
+                }
+            }
+            
+            // Set coupon discount
+            $new_order->setDiscounts();
+            
+            if (count($errors) > 0) {
+                return redirect('admin/orders')->withErrors($errors);
+            }
         }
         
         return redirect('admin/orders');