+++ src/database/migrations/2014_05_15_070522_create_pricelists_table.php
@@ -12,17 +12,21 @@ class CreatePricelistsTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('pricelists', function(Blueprint $table)
-		{
-			$table->increments('id');
-            $table->decimal('price', 8, 2)->default(0);
-			$table->integer('module_id')->unsigned();
-			$table->foreign('module_id')->references('id')->on('modules');
-			$table->integer('membership_id')->unsigned();
-			$table->foreign('membership_id')->references('id')->on('memberships');
-			$table->unique(array('module_id', 'membership_id'));
-			$table->timestamps();
-		});
+        if (! Schema::hasTable('pricelists')) {
+            Schema::create('pricelists', function(Blueprint $table)
+            {
+                $table->increments('id');
+                $table->decimal('price', 8, 2)->default(0);
+                $table->integer('module_id')->unsigned();
+                $table->foreign('module_id')->references('id')->on('modules');
+                $table->integer('membership_id')->unsigned();
+                $table->foreign('membership_id')->references('id')->on('memberships');
+                $table->unique(array('module_id', 'membership_id'));
+                $table->timestamps();
+                // Need to use InnoDB to support foreign key
+                $table->engine = 'InnoDB';
+            });
+        }
 	}
 
 	/**