+++ src/database/migrations/2015_08_18_000000_drop_coupon_user_table.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class DropCouponUserTable extends Migration {
+
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+        Schema::dropIfExists('coupon_user');
+	}
+
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+        if (! Schema::hasTable('coupon_user')) {
+            Schema::create('coupon_user', function(Blueprint $table)
+            {
+                $table->increments('id');
+                $table->integer('coupon_id')->unsigned();
+                $table->integer('user_id')->unsigned();
+                $table->foreign('coupon_id')->references('id')->on('coupons')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
+	}
+
+}